pdfUtils/requirements.txt
flake8==3.8.2
importlib-metadata==1.6.1
mccabe==0.6.1
Pillow==7.1.2
pycodestyle==2.6.0
pyflakes==2.2.0
zipp==3.1.0
